Refactor #349 - move Timestamp provider to data module
