+++ core/data-di/build.gradle.kts
@@ -10,6 +10,7 @@ android {
 dependencies {
     api(projects.core.data)
     implementation(projects.core.dataImpl)
+    implementation(projects.core.dataPlatformImpl)
     implementation(projects.core.localDi)
     implementation(projects.core.network)
