+++ core/data-platform-impl/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    alias(libs.plugins.nlab.android.library)
+}
+
+android {
+    namespace = "com.nlab.reminder.core.data.platform.impl"
+}
+
+dependencies {
+    implementation(projects.core.data)
+}
