+++ core/data-platform-impl/src/main/AndroidManifest.xml
