--- core/foundation-di/build.gradle.kts
@@ -1,12 +1,0 @@
-plugins {
-    alias(libs.plugins.nlab.android.library)
-    alias(libs.plugins.nlab.android.library.inject)
-}
-
-android {
-    namespace = "com.nlab.reminder.core.foundation.di"
-}
-
-dependencies {
-    api(projects.core.foundation)
-}
