--- core/foundation-impl/build.gradle.kts
@@ -1,11 +1,0 @@
-plugins {
-    alias(libs.plugins.nlab.android.library)
-}
-
-android {
-    namespace = "com.nlab.reminder.core.foundation.impl"
-}
-
-dependencies {
-    implementation(projects.core.foundation)
-}
