+++ settings.gradle.kts
@@ -63,13 +63,12 @@ include(
     ":core:data",
     ":core:data-di",
     ":core:data-impl",
+    ":core:data-platform-impl",
     ":core:data-test",
     ":core:designsystem",
     ":core:domain",
     ":core:domain-di",
     ":core:foundation",
-    ":core:foundation-di",
-    ":core:foundation-impl",
     ":core:inject",
     ":core:kotlin",
     ":core:kotlin-test",
